FW: create sandstone_chrono.{h,cpp}
Move all the time-related functions there.
