+++ framework/sysdeps/linux/effective_cpu_freq.hpp
@@ -14,8 +14,6 @@
 #include <x86intrin.h>
 #include "sandstone_p.h"
 
-extern int get_monotonic_time_now(struct timespec *);
-
 class CPUTimeFreqStamp
 {
 public:
